Various warning fixes
2004-04-18  Nektarios Papadopoulos <npapadop@inaccessnetworks.com>

        * kaffe/kaffevm/classMethod.c,
        kaffe/kaffevm/classMethod.h,
        kaffe/kaffevm/debug.c,
        kaffe/kaffevm/external.c,
        kaffe/kaffevm/file.h,
        kaffe/kaffevm/findInJar.c,
        kaffe/kaffevm/fp.c,
        kaffe/kaffevm/jar.c,
        kaffe/kaffevm/jni.c,
        kaffe/kaffevm/object.c,
        kaffe/kaffevm/object.h,
        kaffe/kaffevm/readClass.c,
        kaffe/kaffevm/soft.c,
        kaffe/kaffevm/stackTrace.c,
        kaffe/kaffevm/string.c,
        kaffe/kaffevm/stringParsing.c,
        kaffe/kaffevm/support.c,
        kaffe/kaffevm/utf8const.c,
        kaffe/kaffevm/verify.c,
        kaffe/kaffevm/mem/gc-incremental.c,
        kaffe/kaffevm/systems/unix-jthreads/syscalls.c,
        libraries/clib/io/AudioPlayer.c,
        libraries/clib/io/File.c,
        libraries/clib/management/Debug.c,
        libraries/clib/management/JIT.c,
        libraries/clib/management/XProfiler.c,
        libraries/clib/native/Array.c,
        libraries/clib/native/Arrays.c,
        libraries/clib/native/ByteToCharIconv.c,
        libraries/clib/native/CharToByteIconv.c,
        libraries/clib/native/Class.c,
        libraries/clib/native/ClassLoader.c,
        libraries/clib/native/Compiler.c,
        libraries/clib/native/MemoryAdvice.c,
        libraries/clib/native/Method.c,
        libraries/clib/native/PrimordialClassLoader.c,
        libraries/clib/native/Runtime.c,
        libraries/clib/native/String.c,
        libraries/clib/native/System.c,
        libraries/clib/native/Thread.c,
        libraries/clib/net/InetAddressImpl.c,
        libraries/clib/net/PlainDatagramSocketImpl.c,
        libraries/clib/net/PlainSocketImpl.c,
        libraries/clib/security/digest.c,
        libraries/clib/zip/Adler32.c,
        libraries/clib/zip/CRC32.c,
        libraries/clib/zip/Deflater.c,
        libraries/clib/zip/Inflater.c:
        Fixed various warnings about unused parameters and
        signedness of variables.

+++ kaffe/kaffevm/external.c
@@ -143,7 +143,7 @@ initNative(void)
 	char* lpath;
 	char* nptr;
 	char* ptr;
-	int len;
+	unsigned int len;
 
 	DBG(INIT, dprintf("initNative()\n"); )
 
@@ -185,7 +185,7 @@ initNative(void)
 			continue;
 		}
 		else {
-			strncpy(lib, ptr, nptr - ptr);
+			strncpy(lib, ptr, (size_t)(nptr - ptr));
 			lib[nptr-ptr] = '\0';
 			nptr += strlen(path_separator);
 		}
